Separate Kaltura modal into Upload and Log tabs
