+++ templates/vue/src/components/modals/KalturaModal.vue
@@ -6,190 +6,22 @@
     size="lg"
     title="Kaltura Sync"
     scrollable
+    body-class="p-0"
     @hidden="$emit('close')"
   >
-    <b-container fluid class="px-1">
-      <b-alert
-        dismissible
-        class="mt-2 sticky-top"
-        variant="danger"
-        :show="!!apiError"
-        @dismissed="apiError = null"
-      >
-        Error: {{ apiError }}
-      </b-alert>
-      <b-overlay :show="videosUploading" variant="white">
-        <template #overlay>
-          <div v-show="!isLatestTapestry" class="different-tapestry-notice">
-            Please wait: videos are currently being uploaded from another Tapestry.
-          </div>
-        </template>
-        <div>Upload Videos to Kaltura</div>
-        <b-form-text>
-          Select videos in this Tapestry that you would like to upload to Kaltura.
-          Only videos added as file uploads can be uploaded.
-        </b-form-text>
-        <b-table
-          ref="videoTable"
-          v-model="allVideos"
-          class="my-3"
-          selectable
-          no-select-on-click
-          show-empty
-          empty-text="There are no videos to upload in this Tapestry."
-          selected-variant=""
-          primary-key="nodeID"
-          :fields="['selected', 'nodeID', 'nodeTitle']"
-          :items="getVideosToUpload"
-          @row-selected="handleVideoSelected"
+    <b-container fluid class="px-0">
+      <b-tabs card>
+        <b-tab
+          title="Upload"
+          :active="tab === 'upload'"
+          @click="changeTab('upload')"
         >
-          <template #head(selected)="{clearSelected}">
-            <b-form-checkbox
-              aria-label="Select all videos"
-              :checked="allVideosSelected"
-              :disabled="videosUploading"
-              @change="$event ? selectAllVideos() : clearSelected()"
-            ></b-form-checkbox>
-          </template>
-          <template #cell(selected)="{rowSelected, selectRow, unselectRow, item}">
-            <b-form-checkbox
-              :aria-label="`Select node ${item.nodeID}, ${item.nodeTitle}`"
-              :checked="rowSelected"
-              :disabled="videosUploading || !item.withinSizeLimit"
-              class="d-inline"
-              @change="$event ? selectRow() : unselectRow()"
-            ></b-form-checkbox>
-            <i
-              v-if="!item.withinSizeLimit"
-              :id="`video-select-info-${item.nodeID}`"
-              class="far fa-question-circle"
-            ></i>
-            <b-tooltip
-              v-if="!item.withinSizeLimit"
-              :target="`video-select-info-${item.nodeID}`"
-              triggers="hover"
-            >
-              This video is too big to be uploaded to Kaltura.
-            </b-tooltip>
-          </template>
-        </b-table>
-        <b-form-group>
-          <b-form-checkbox v-model="useKalturaPlayer" :disabled="videosUploading">
-            Switch uploaded videos to use Kaltura media player
-          </b-form-checkbox>
-        </b-form-group>
-      </b-overlay>
-      <b-button
-        id="start-upload-button"
-        block
-        variant="light"
-        :class="startButtonDisabled ? 'disabled' : ''"
-        :disabled="startButtonDisabled"
-        @click="startVideoUpload"
-      >
-        <b-spinner v-if="videosUploading && isLatestTapestry" small></b-spinner>
-        <div :style="startButtonDisabled ? 'opacity: 50%;' : ''">
-          Start Upload
-        </div>
-      </b-button>
-      <b-button
-        id="stop-upload-button"
-        block
-        variant="light"
-        :class="stopButtonDisabled ? 'disabled' : ''"
-        :disabled="stopButtonDisabled"
-        @click="requestStopVideoUpload"
-      >
-        <div :style="stopButtonDisabled ? 'opacity: 50%;' : ''">
-          Stop Upload
-        </div>
-      </b-button>
-      <b-alert class="mt-2" :show="hasRequestedStop" variant="success">
-        Successfully canceled the upload. Note: Videos already being processed will
-        still be uploaded to Kaltura, but no more videos will be started. Please be
-        patient as processing these videos could take some time.
-      </b-alert>
-      <div class="mt-3">Upload Log</div>
-      <b-form-text>
-        Automatically refreshes every 15 seconds; you can also manually refresh.
-        <br />
-        Once the upload has completed, please reload the page to see the updated node
-        content.
-      </b-form-text>
-      <b-alert
-        class="mt-2"
-        dismissible
-        :show="uploadError && isLatestTapestry"
-        variant="danger"
-        @dismissed="clearUploadError"
-      >
-        <p>
-          The upload did not complete due to an error in the server.
-        </p>
-        <p class="mb-1">
-          If any videos are still Converting, to avoid re-uploading them, we
-          recommend running
-          <b>Clean Uploaded Videos</b>
-          under the WordPress Settings > Tapestry.
-        </p>
-      </b-alert>
-      <div class="my-3 d-flex flex-row align-items-center">
-        <b-form-select v-model="perPage" class="per-page-select mr-2">
-          <b-form-select-option :value="10">10 per page</b-form-select-option>
-          <b-form-select-option :value="25">25 per page</b-form-select-option>
-          <b-form-select-option :value="50">50 per page</b-form-select-option>
-          <b-form-select-option :value="100">100 per page</b-form-select-option>
-          <b-form-select-option :value="0">Show all entries</b-form-select-option>
-        </b-form-select>
-        <b-pagination
-          v-model="currentPage"
-          class="mb-0"
-          aria-controls="upload-log-table"
-          :total-rows="uploadLogLength"
-          :per-page="perPage"
-        ></b-pagination>
-        <b-button size="sm" class="ml-auto" @click="refreshVideoUploadStatus">
-          Refresh
-        </b-button>
-      </div>
-      <b-table
-        id="upload-log-table"
-        ref="uploadStatusTable"
-        class="my-3"
-        responsive
-        striped
-        show-empty
-        empty-text="No videos were uploaded from this Tapestry."
-        :fields="[
-          {
-            key: 'uploadTime',
-            class: 'align-top',
-          },
-          {
-            key: 'nodeID',
-            class: 'align-top',
-          },
-          {
-            key: 'nodeTitle',
-            class: 'align-top',
-          },
-          {
-            key: 'uploadStatus',
-            class: 'align-top',
-          },
-          {
-            key: 'kalturaID',
-            class: 'align-top',
-          },
-          {
-            key: 'additionalInfo',
-            class: 'align-top',
-          },
-        ]"
-        :items="getUploadLog"
-        :current-page="currentPage"
-        :per-page="perPage"
-      ></b-table>
+          <kaltura-upload-tab ref="uploadTab" @upload-start="handleUploadStart" />
+        </b-tab>
+        <b-tab title="Log" :active="tab === 'log'" @click="changeTab('log')">
+          <kaltura-upload-log-tab ref="logTab" @refresh="refresh" />
+        </b-tab>
+      </b-tabs>
     </b-container>
     <template slot="modal-footer">
       <b-button size="sm" variant="secondary" @click="closeModal">
@@ -200,11 +32,15 @@
 </template>
 
 <script>
-import client from "@/services/TapestryAPI"
-import { data as wpData } from "@/services/wp"
-import ErrorHelper from "@/utils/errorHelper"
+import { names } from "@/config/routes"
+import KalturaUploadTab from "./KalturaUploadTab"
+import KalturaUploadLogTab from "./KalturaUploadLogTab"
 
 export default {
+  components: {
+    KalturaUploadTab,
+    KalturaUploadLogTab,
+  },
   props: {
     show: {
       type: Boolean,
@@ -213,186 +49,70 @@ export default {
   },
   data() {
     return {
-      videosUploading: false,
-      uploadError: false,
-      apiError: null,
-      useKalturaPlayer: false,
-      allVideos: [],
-      selectedVideos: [],
-      uploadStatusRefreshTimer: 0,
-      hasRequestedStop: false,
-      latestTapestryID: "",
-      uploadLogLength: 0,
-      currentPage: 1,
-      perPage: 10,
+      refreshTimer: 0,
     }
   },
   computed: {
-    allVideosSelected() {
-      return (
-        this.allUploadableVideos.length === this.selectedVideos.length &&
-        this.allUploadableVideos.length > 0
-      )
-    },
-    allUploadableVideos() {
-      return this.allVideos.filter(video => video.withinSizeLimit)
-    },
-    canStartUpload() {
-      return this.selectedVideos.length > 0
-    },
-    isLatestTapestry() {
-      return this.latestTapestryID === wpData.postId
-    },
-    startButtonDisabled() {
-      return this.videosUploading || !this.canStartUpload
-    },
-    stopButtonDisabled() {
-      return !this.videosUploading || !this.isLatestTapestry
+    tab() {
+      return this.$route.params.tab
     },
   },
   watch: {
-    videosUploading(inProgress) {
-      if (!inProgress) {
-        this.refreshVideosToUpload()
-      }
+    tab: {
+      immediate: true,
+      handler(requestedTab) {
+        if (this.show) {
+          const acceptedTabs = ["upload", "log"]
+          if (!acceptedTabs.includes(requestedTab)) {
+            this.$router.replace({
+              name: names.KALTURAMODAL,
+              params: { nodeId: this.$route.params.nodeId, tab: "upload" },
+              query: this.$route.query,
+            })
+          }
+        }
+      },
     },
     show: {
       immediate: true,
       handler(show) {
         if (show) {
-          this.selectedVideos = []
           setTimeout(() => {
-            this.refreshVideosToUpload()
+            this.refresh()
           }, 0)
 
-          this.uploadStatusRefreshTimer = setInterval(
-            this.refreshVideoUploadStatus,
-            15 * 1000
-          )
-          setTimeout(() => {
-            this.refreshVideoUploadStatus()
-          }, 0)
+          this.refreshTimer = setInterval(this.refresh, 15 * 1000)
         } else {
-          this.cancelUploadStatusRefresh()
+          this.cancelRefresh()
         }
       },
     },
   },
   beforeDestroy() {
-    this.cancelUploadStatusRefresh()
+    this.cancelRefresh()
   },
   methods: {
     closeModal() {
       this.$root.$emit("bv::hide::modal", "kaltura-modal")
     },
-    selectAllVideos() {
-      for (let i = 0; i < this.allVideos.length; i++) {
-        if (this.allVideos[i].withinSizeLimit) {
-          this.$refs.videoTable.selectRow(i)
-        }
-      }
-    },
-    handleVideoSelected(rows) {
-      this.selectedVideos = rows
-    },
-    cancelUploadStatusRefresh() {
-      clearInterval(this.uploadStatusRefreshTimer)
-      this.uploadStatusRefreshTimer = 0
-    },
-    startVideoUpload() {
-      this.videosUploading = true
-      this.latestTapestryID = wpData.postId
-
-      client
-        .startKalturaUpload(
-          this.selectedVideos.map(video => video.nodeID),
-          this.useKalturaPlayer
-        )
-        .catch(error => {
-          // Kaltura availability changed unexpectedly
-          this.apiError = ErrorHelper.getErrorMessage(error)
-        })
-
-      setTimeout(this.refreshVideoUploadStatus, 500)
-    },
-    async requestStopVideoUpload() {
-      await client.requestStopKalturaUpload()
-      this.hasRequestedStop = true
-      setTimeout(() => {
-        this.hasRequestedStop = false
-      }, 10 * 1000)
-    },
-    getVideosToUpload(ctx, callback) {
-      client
-        .getVideosToUpload()
-        .then(data => {
-          callback(data)
-        })
-        .catch(() => {
-          callback([])
-        })
-
-      return null
-    },
-    getUploadLog(ctx, callback) {
-      client
-        .getKalturaUploadLog(ctx.currentPage, ctx.perPage)
-        .then(data => {
-          callback(data.videos)
-          this.uploadLogLength = data.totalCount
-        })
-        .catch(() => {
-          callback([])
-        })
-
-      return null
-    },
-    refreshVideosToUpload() {
-      this.$refs.videoTable.refresh()
-    },
-    refreshVideoUploadStatus() {
-      this.$refs.uploadStatusTable.refresh()
-      client.getKalturaUploadStatus().then(data => {
-        this.videosUploading = data.inProgress
-        this.uploadError = data.error
-        this.latestTapestryID = data.latestTapestryID
+    changeTab(tab) {
+      this.$router.push({
+        name: names.KALTURAMODAL,
+        params: { nodeId: this.$route.params.nodeId, tab },
+        query: this.$route.query,
       })
     },
-    async clearUploadError() {
-      this.uploadError = false
-      await client.clearKalturaUploadError()
+    cancelRefresh() {
+      clearInterval(this.refreshTimer)
+      this.refreshTimer = 0
+    },
+    handleUploadStart() {
+      setTimeout(this.refresh, 1000)
+    },
+    refresh() {
+      this.$refs.uploadTab.refresh()
+      this.$refs.logTab.refresh()
     },
   },
 }
 </script>
-
-<style lang="scss" scoped>
-#start-upload-button {
-  position: relative;
-  > span {
-    position: absolute;
-    height: 1.5em;
-    width: 1.5em;
-  }
-}
-
-#start-upload-button,
-#stop-upload-button {
-  &.disabled {
-    pointer-events: none;
-    cursor: not-allowed;
-  }
-}
-
-.different-tapestry-notice {
-  text-align: center;
-  background: #fff;
-  border-radius: 0.5em;
-  padding: 1em;
-  max-width: 500px;
-}
-
-.per-page-select {
-  width: inherit !important;
-}
-</style>
